Renamed keymappings.lua and updated plugins.lua
