+++ lua/plugins.lua
@@ -1,7 +1,19 @@
 --| @plugins.lua |--
 
--- check if packer is installed (~/local/share/nvim/site/pack)
-local packer_exists = pcall(vim.cmd, [[packadd packer.nvim]])
+
+-- installs packer if not already installed
+
+local execute = vim.api.nvim_command
+local fn = vim.fn
+
+local install_path = fn.stdpath("data") .. "/site/pack/packer/start/packer.nvim"
+
+if fn.empty(fn.glob(install_path)) > 0 then
+    execute("!git clone https://github.com/wbthomason/packer.nvim " .. install_path)
+    execute "packadd packer.nvim"
+end
+
+-- plugins list
 
 return require('packer').startup( function() 
   use 'wbthomason/packer.nvim'			-- package manager
